+++ src/Components/Details/ShiftCheckbox/ShiftCheckbox.js
@@ -2,7 +2,7 @@ import { useTranslation } from 'react-i18next';
 import '../../../styles/shift-checkbox.scss';
 import formatTime from '../../../Utils/formatTime.js';
 
-const ShiftCheckbox = ({ id, startTime, endTime, numVolunteers, maxVolunteers, onChange }) => {
+const ShiftCheckbox = ({ startTime, endTime, numVolunteers, maxVolunteers, onChange }) => {
 
     const { t } = useTranslation();
 
@@ -16,7 +16,7 @@ const ShiftCheckbox = ({ id, startTime, endTime, numVolunteers, maxVolunteers, o
             <div className="shift_checkbox">
                 <label htmlFor="shift_hours">
                     <input type='checkbox' id="shift_hours" onChange={handleShiftChange}/>
-                    {id}: {formatTime(startTime)} - {formatTime(endTime)}
+                    {formatTime(startTime)} - {formatTime(endTime)}
                 </label>
             </div>
             <p id="shift_num_of_volunteers">
