+++ src/Utils/tests/formatTime.test.js
@@ -8,4 +8,12 @@ describe('formatTime', () => {
     const formattedTime = formatTime(inputTime);
     expect(formattedTime).toEqual(expectedOutput);
   });
+
+  it('format a time with seconds that are not 00', () => {
+    const inputTime = '01:02:03';
+    const expectedOutput = '01:02';
+
+    const formattedTime = formatTime(inputTime);
+    expect(formattedTime).toEqual(expectedOutput);
+  });
 });
